+++ src/test/java/AutoFilePathsTest.java
@@ -1,0 +1,1 @@
+
